+++ vm/src/hint_processor/cairo_1_hint_processor/hint_processor.rs
@@ -474,10 +474,10 @@ impl Cairo1HintProcessor {
         remainder1: &CellRef,
     ) -> Result<(), HintError> {
         let pow_2_128 = BigUint::from(u128::MAX) + 1u32;
-        let dividend0 = get_val(vm, dividend0)?.to_biguint();
-        let dividend1 = get_val(vm, dividend1)?.to_biguint();
-        let divisor0 = get_val(vm, divisor0)?.to_biguint();
-        let divisor1 = get_val(vm, divisor1)?.to_biguint();
+        let dividend0 = res_operand_get_val(vm, dividend0)?.to_biguint();
+        let dividend1 = res_operand_get_val(vm, dividend1)?.to_biguint();
+        let divisor0 = res_operand_get_val(vm, divisor0)?.to_biguint();
+        let divisor1 = res_operand_get_val(vm, divisor1)?.to_biguint();
         let dividend: BigUint = dividend0 + dividend1.shl(128);
         let divisor = divisor0 + divisor1.shl(128);
         let (quotient, remainder) = dividend.div_rem(&divisor);
@@ -1155,10 +1155,10 @@ impl Cairo1HintProcessor {
         t_or_k1: &CellRef,
     ) -> Result<(), HintError> {
         let pow_2_128 = BigInt::from(u128::MAX) + 1u32;
-        let b0 = get_val(vm, b0)?.to_bigint();
-        let b1 = get_val(vm, b1)?.to_bigint();
-        let n0 = get_val(vm, n0)?.to_bigint();
-        let n1 = get_val(vm, n1)?.to_bigint();
+        let b0 = res_operand_get_val(vm, b0)?.to_bigint();
+        let b1 = res_operand_get_val(vm, b1)?.to_bigint();
+        let n0 = res_operand_get_val(vm, n0)?.to_bigint();
+        let n1 = res_operand_get_val(vm, n1)?.to_bigint();
         let b: BigInt = b0.clone() + b1.clone().shl(128);
         let n: BigInt = n0 + n1.shl(128);
         let ExtendedGcd {
@@ -1217,14 +1217,14 @@ impl Cairo1HintProcessor {
         n_mul_mods: &ResOperand,
         mul_mod_builtin_ptr: &ResOperand,
     ) -> Result<(), HintError> {
-        let n_add_mods = get_val(vm, n_add_mods)?;
+        let n_add_mods = res_operand_get_val(vm, n_add_mods)?;
         let n_add_mods =
             n_add_mods
                 .to_usize()
                 .ok_or(HintError::Math(MathError::Felt252ToUsizeConversion(
                     Box::from(n_add_mods),
                 )))?;
-        let n_mul_mods = get_val(vm, n_mul_mods)?;
+        let n_mul_mods = res_operand_get_val(vm, n_mul_mods)?;
         let n_mul_mods =
             n_mul_mods
                 .to_usize()
@@ -1263,6 +1263,8 @@ impl HintProcessorLogic for Cairo1HintProcessor {
         _reference_ids: &HashMap<String, usize>,
         //List of all references (key corresponds to element of the previous dictionary)
         _references: &[HintReference],
+        // List of accessible scopes in the hint
+        _accessible_scopes: &[String],
     ) -> Result<Box<dyn Any>, VirtualMachineError> {
         let data = hint_code.parse().ok().and_then(|x: usize| self.hints.get(&x).cloned())
         .ok_or_else(|| VirtualMachineError::CompileHintFail(